Android App에 무작정 아무 GATT SVC에서 Nofify하려고 해서 GAP에 NOTIFY를 넣어서 특정 CHAR를 선택할 수 있도록 땜질함

+++ main/include/heart_rate.h
@@ -14,7 +14,8 @@
 #define HEART_RATE_TASK_PERIOD (1000 / portTICK_PERIOD_MS)
 
 /* Public function declarations */
-uint8_t get_heart_rate(void);
+//  uint8_t get_heart_rate(void);
+int get_heart_rate(void);
 void update_heart_rate(void);
 
 #endif // HEART_RATE_H
